[tb/update-locations] Rename and update orderer public option utils
