+++ packages/perseus-core/src/index.ts
@@ -37,4 +37,4 @@ export type {ExpressionDefaultWidgetOptions} from "./widgets/expression";
 
 export type * from "./widgets/logic-export.types";
 
-export {default as getOrdererPublicWidgetOptions} from "./utils/orderer-util";
+export {default as getOrdererPublicWidgetOptions} from "./widgets/orderer/orderer-util";
